git-svn-id: https://www.ibr.cs.tu-bs.de/svn/emma/trunk/src/ibrdtn@21983 1cffb75b-d306-6187-5d55-5a277715007a

+++ src/data/Block.cpp
@@ -55,7 +55,7 @@ namespace dtn
 			size_t len = 0;
 			len += writer.write(_blocktype);
 			len += writer.write(_procflags);
-			len += writer.write( _blobref.getSize() );
+			len += writer.write( (u_int64_t)_blobref.getSize() );
 			return len;
 		}
